Altered OrderSummary's functionality.
The OrderSummary class is now meant to be a superclass shared between
all the different windows of the GUI. The base class no longer has
functionality for editing and saving orders, but can only view them.
More functionality can easily be added by subclassing, as proven in the
new ChefOrderSummary and OperatorOrderSummary classes.

The old OrderView and OrderList are both replaced by the new
OrderSummary.

OrderListener is now depricated until its use is discussed
within the group.